Add a custom input layer with an optional template type for input data
The input layer has an option of "data_type." It uses DataType by
default. "int16" can be used.
