+++ src/proto/layers.proto
@@ -323,6 +323,7 @@ message Layer {
     bool data_set_per_model = 1;  // Default: false
     string io_buffer = 2;         // Options: "partitioned" (default)
     string target_mode = 3;       // Options: "classification" (default), "regression", "reconstruction", "N/A"
+    string data_type = 4;         // Options: use DataType if unspecified
   }
 
   //////////////////////
